% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis_sets.R
\name{basis_sets}
\alias{basis_sets}
\alias{legendre_set}
\alias{legendre_M}
\alias{ns_set}
\alias{fourier_M}
\alias{fourier_set}
\title{Basis sets for for function approximation}
\usage{
legendre_set(df = 3, left = -1, right = 1)

legendre_M(x, df, left = -1, right = 1)

ns_set(df = 3, left = -1, right = 1)

fourier_M(x, n, fperiod = NULL)

fourier_set(df, left = -1, right = 1)
}
\arguments{
\item{df}{number of basis functions to construct}

\item{left}{number giving left-hand boundary of the interval}

\item{right}{number giving right-hand boundary of the interval}

\item{fperiod}{number giving the fundamental period length for the Fourier basis}
}
\value{
The \verb{_M} functions return a model matrix. The \verb{_set} functions
return a function with arguments \code{x} and \code{n}. The integer \code{n} specifies
which function to use, while \code{x} is the set of values at which to evaluate
that function.
}
\description{
These functions generate the mathematical functions for three
different basis sets: Fourier (sines), Legendre (orthogonal polynomials),
and Splines (low-order smooth approximation)
}
\details{
For each basis, there are two different forms for the
generating functions. Names ending in \verb{_set} create a set of functions with
arguments \code{x} and \code{n}, where integer \code{n} provides an index into the set.
The same names with a \verb{_M} suffix produce a model matrix
corresponding to a specified set of x values. These are useful
with \code{lm()} and similar model-building functions in the same way that
\code{poly()} and \code{ns()} are useful. (\code{ns_M()} is just an alias for \code{splines::ns()}.) Like
\code{poly()} and \code{ns()}, the \verb{_M} suffix functions do \emph{NOT} include an
intercept column.
}
