% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigRadioButtons.R
\name{bigRadioButtons}
\alias{bigRadioButtons}
\alias{bigRadioButtonItem}
\title{Radio buttons that display html contents.}
\usage{
bigRadioButtons(id, label, choices, origNum)

bigRadioButtonItem(id, value, string)
}
\arguments{
\item{id}{the same id given to \code{bigRadioButtons}}

\item{label}{character string label}

\item{choices}{a character vector, one for each choice}

\item{origNum}{something to do with randomization?}

\item{value}{identifier for the individual radio button, typically, 1, 2, 3, ...
for successive buttons.}

\item{string}{the HTML string in the \verb{<div>} for the button.}
}
\description{
Creates a radio-button input for Shiny where the item labels are arbitrary
html. Thus, text, MathJax, images, ... can be included in the label.
}
