% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strings.R
\docType{data}
\name{Cello}
\alias{Cello}
\title{Recordings of musical notes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 62208 rows and 2 columns.
}
\usage{
Cello
}
\description{
Cello and violin notes. Data frames with \code{t} and \code{y} columns
}
\details{
\code{Cello} and \code{Violin} are complete notes

\code{Cello_seg} and \code{Violin_seg} are extracts, approximately 90 ms long

Source: https://philharmonia.co.uk/resources/sound-samples/
}
\keyword{datasets}
