% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{curvature}
\alias{curvature}
\title{Transform a function into it's curvature function}
\usage{
curvature(ftilde)
}
\arguments{
\item{ftilde}{A tilde expression describing the function and the
variable with respect to which the curvature should be computed.}
}
\value{
A function with the same argument as that described by the
tilde expression argument.
}
\description{
curvature works much like \code{D()}, but instead of returning a function
containing the first or second derivative, the function instead
gives the curvature of the parent function.
}
\details{
Works only for functions of a single variable.

Curvature is 1/radius of the tangent circle that can be inscribed
in the function at any given input.
}
\examples{
cfun <- curvature(sin(x) ~ x)

}
