% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driving.R
\name{heading}
\alias{heading}
\alias{slope}
\alias{speed}
\alias{mpg}
\title{Functions for the "Driving with integrals" project}
\usage{
heading(t)

slope(t)

speed(t)

mpg(t)
}
\arguments{
\item{t}{Time from 0 to 60 minutes.}
}
\description{
Generate speed, fuel economy, and heading data
}
