% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics-examples.R
\name{dynmodels}
\alias{dynmodels}
\alias{dt_rabbit}
\alias{dt_fox}
\title{Examples of dynamical functions}
\usage{
dt_rabbit(r, f, alpha = 0.66, beta = 1.33)

dt_fox(r, f, delta = -1, beta = 1)
}
\arguments{
\item{r, }{f, V, ... state variables.}

\item{alpha, }{beta, gamma, delta, ... parameters in the dynamical functions.}
}
\description{
These are implementations of the dynamical functions for examples
shown in the text.
\itemize{
\item \code{dt_rabbit()} and \code{dt_fox()} are for the rabbit/fox system
}
}
