% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter_e.R
\docType{data}
\name{letter.e}
\alias{letter.e}
\title{Knot points describing the shape of lowercase "e"}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 3 columns.
}
\usage{
letter.e
}
\description{
Knot points describing the shape of lowercase "e"
}
\keyword{datasets}
