% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_task.R
\name{match_assignment}
\alias{match_assignment}
\title{Check a mosaic programming task}
\usage{
match_assignment(target, new)
}
\arguments{
\item{target}{character string containing the CORRECT expression}

\item{new}{character string containing candidate for checking against target
#export}
}
\description{
Utilities for the command trainer
}
