% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QALY.R
\docType{data}
\name{QALY_impact}
\alias{QALY_impact}
\title{Quality Adjusted Life Years}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 3 columns.
}
\usage{
QALY_impact
}
\description{
A small, made-up data set used in an exercise about optimizing
health-care expenditures.
}
\keyword{datasets}
