% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Flexibly specify a domain for plotting}
\usage{
region(...)
}
\description{
Domains are used so often for plotting that it's nice
to have a more compact syntax. Within \code{region()} you
can replace c(left, right) with \code{left | right}, \code{left ^ right},
and so on. (Typically, left and right are numerals, as in \code{0 | 10}.)
}
