% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robot.stations.R
\docType{data}
\name{Robot.stations}
\alias{Robot.stations}
\title{Waypoints on a path in two dimensions used as a textbook example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 3 columns.
}
\usage{
Robot.stations
}
\description{
A data frame with \code{x} and \code{y} components
}
\keyword{datasets}
