% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSA_2007.R
\docType{data}
\name{SSA_2007}
\alias{SSA_2007}
\title{US Mortality table from 2007}
\format{
An object of class \code{data.frame} with 240 rows and 4 columns.
}
\usage{
SSA_2007
}
\description{
US Mortality table from 2007
}
\keyword{datasets}
