% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vowels.R
\docType{data}
\name{Vowel_ee}
\alias{Vowel_ee}
\title{Recordings of vowel sounds}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 859 rows and 2 columns.
}
\usage{
Vowel_ee
}
\description{
"Oh" refers to "o" as in "stone." "Ee" to "eel".
}
\details{
\itemize{
\item \code{Oh_sound} and \code{Ee_sound} contain the complete utterances.
\item \code{Vowel_ee} and \code{Vowel_oh} are snippets of about 0.04 seconds.
}
}
\keyword{datasets}
