% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture}
\alias{capture}
\title{Displays the result of a command with the command itself}
\usage{
capture(cmd, message = "", digits = 4)
}
\arguments{
\item{cmd}{An R expression, typically piped into \code{capture()}}

\item{message}{a comment to print after the result of cmd.}
}
\description{
In \code{{learnr}}, the results of running an exercise are posted below the code-entry
box. When there are multiple commands, this breaks the direct link between the command
itself and the printed results. \code{capture()} prints out both the command and the result
along with an optional message. \code{annot()} turns a character string into a line
that looks like a comment.
}
\details{
You \strong{must} use the built-in pipe (\verb{|>}) into \code{capture()}. Use whatever
you like to the left of \verb{|>}.
Meant for results that are very short vectors, either numeric or character string.
}
\examples{
cos(7) |> capture("Example 3")
annot("Show the plot")
}
